Crate linear_map

A map implemented by searching linearly in a vector.

See the LinearMap type for details.

Modules

serde

An optional implementation of serialization/deserialization. Reference implementations used:

set

A set implemented by searching linearly in a vector.

Macros

linear_map

Creates a LinearMap from a list of key-value pairs.

Structs

Drain

A draining iterator over a LinearMap.

IntoIter

A consuming iterator over a LinearMap.

Iter

An iterator yielding references to a LinearMap’s keys and their corresponding values.

IterMut

An iterator yielding references to a LinearMap’s keys and mutable references to their corresponding values.

Keys

An iterator yielding references to a LinearMap’s keys in arbitrary order.

LinearMap

A map implemented by searching linearly in a vector.

OccupiedEntry

A view into a single occupied location in a LinearMap.

VacantEntry

A view into a single vacant location in a LinearMap.

Values

An iterator yielding references to a LinearMap’s values in arbitrary order.

Enums

Entry

A view into a single entry in a LinearMap.