Struct linear_map::OccupiedEntry

pub struct OccupiedEntry<'a, K: 'a, V: 'a> {
    // some fields omitted
}

A view into a single occupied location in a LinearMap.

Methods

impl<'a, K, V> OccupiedEntry<'a, K, V>

fn get(&self) -> &V

Gets a reference to the value in the entry.

fn get_mut(&mut self) -> &mut V

Gets a mutable reference to the value in the entry.

fn into_mut(self) -> &'a mut V

Converts the OccupiedEntry into a mutable reference to the value in the entry with a lifetime bound to the map itself

fn insert(&mut self, value: V) -> V

Sets the value of the entry, and returns the entry's old value

fn remove(self) -> V

Takes the value out of the entry, and returns it