Struct linear_map::VacantEntry

pub struct VacantEntry<'a, K: 'a, V: 'a> {
    // some fields omitted
}

A view into a single empty location in a LinearMap.

Methods

impl<'a, K, V> VacantEntry<'a, K, V>

fn insert(self, value: V) -> &'a mut V

Sets the value of the entry with the VacantEntry's key, and returns a mutable reference to it