Crate linear_map

A module providing a map implementation LinearMap backed by a vector.

Structs

IntoIter
A consuming iterator over a map.

Iter
The iterator returned by LinearMap::iter.

IterMut
The iterator returned by LinearMap::iter_mut.

Keys
The iterator returned by LinearMap::keys.

LinearMap
A very simple map implementation backed by a vector.

OccupiedEntry
A view into a single occupied location in a LinearMap.

VacantEntry
A view into a single empty location in a LinearMap.

Values
The iterator returned by LinearMap::values.

Enums

Entry
A view into a single location in a map, which may be vacant or occupied.