Struct linear_map::set::LinearSet

pub struct LinearSet<T> { /* fields omitted */ }

An implementation of a set using the underlying representation of a LinearMap where the value is ().

Examples

use linear_map::set::LinearSet;;
// Type inference lets us omit an explicit type signature (which
// would be `LinearSet<&str>` in this example).
let mut books = LinearSet::new();

// Add some books.
books.insert("A Dance With Dragons");
books.insert("To Kill a Mockingbird");
books.insert("The Odyssey");
books.insert("The Great Gatsby");

// Check for a specific one.
if !books.contains("The Winds of Winter") {
    println!("We have {} books, but The Winds of Winter ain't one.",
             books.len());
}

// Remove a book.
books.remove("The Odyssey");

// Iterate over everything.
for book in &books {
    println!("{}", book);
}

The easiest way to use LinearSet with a custom type is to derive Eq. We must also derive PartialEq, this will in the future be implied by Eq.

use linear_map::set::LinearSet;;
#[derive(Eq, PartialEq, Debug)]
struct Viking<'a> {
    name: &'a str,
    power: usize,
}

let mut vikings = LinearSet::new();

vikings.insert(Viking { name: "Einar", power: 9 });
vikings.insert(Viking { name: "Einar", power: 9 });
vikings.insert(Viking { name: "Olaf", power: 4 });
vikings.insert(Viking { name: "Harald", power: 8 });

// Use derived implementation to print the vikings.
for x in &vikings {
    println!("{:?}", x);
}

Methods

impl<T: Eq> LinearSet<T>

fn new() -> LinearSet<T>

Creates an empty LinearSet.

Examples

use linear_map::set::LinearSet;;
let mut set: LinearSet<i32> = LinearSet::new();

fn with_capacity(capacity: usize) -> LinearSet<T>

Creates an empty LinearSet with space for at least n elements in the map.

Examples

use linear_map::set::LinearSet;;
let mut set: LinearSet<i32> = LinearSet::with_capacity(10);

impl<T> LinearSet<T> where
T: Eq,

fn capacity(&self) -> usize

Returns the number of elements the set can hold without reallocating.

Examples

use linear_map::set::LinearSet;;
let set: LinearSet<i32> = LinearSet::with_capacity(100);
assert!(set.capacity() >= 100);

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the LinearSet. The collection may reserve more space to avoid frequent reallocations.

Panics

Panics if the new allocation size overflows usize.

Examples

use linear_map::set::LinearSet;;
let mut set: LinearSet<i32> = LinearSet::new();
set.reserve(10);

fn shrink_to_fit(&mut self)

Shrinks the capacity of the set as much as possible. It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

Examples

use linear_map::set::LinearSet;;

let mut set = LinearSet::with_capacity(100);
set.insert(1);
set.insert(2);
assert!(set.capacity() >= 100);
set.shrink_to_fit();
assert!(set.capacity() >= 2);

fn iter(&self) -> Iter<T>

An iterator visiting all elements in arbitrary order. Iterator element type is &'a T.

Examples

use linear_map::set::LinearSet;;
let mut set = LinearSet::new();
set.insert("a");
set.insert("b");

// Will print in an arbitrary order.
for x in set.iter() {
    println!("{}", x);
}

fn difference<'a>(&'a self, other: &'a LinearSet<T>) -> Difference<'a, T>

Visit the values representing the difference.

Examples

use linear_map::set::LinearSet;;
let a: LinearSet<_> = [1, 2, 3].iter().cloned().collect();
let b: LinearSet<_> = [4, 2, 3, 4].iter().cloned().collect();

// Can be seen as `a - b`.
for x in a.difference(&b) {
    println!("{}", x); // Print 1
}

let diff: LinearSet<_> = a.difference(&b).cloned().collect();
assert_eq!(diff, [1].iter().cloned().collect());

// Note that difference is not symmetric,
// and `b - a` means something else:
let diff: LinearSet<_> = b.difference(&a).cloned().collect();
assert_eq!(diff, [4].iter().cloned().collect());

fn symmetric_difference<'a>(
&'a self,
other: &'a LinearSet<T>
) -> SymmetricDifference<'a, T>

Visit the values representing the symmetric difference.

Examples

use linear_map::set::LinearSet;;
let a: LinearSet<_> = [1, 2, 3].iter().cloned().collect();
let b: LinearSet<_> = [4, 2, 3, 4].iter().cloned().collect();

// Print 1, 4 in arbitrary order.
for x in a.symmetric_difference(&b) {
    println!("{}", x);
}

let diff1: LinearSet<_> = a.symmetric_difference(&b).cloned().collect();
let diff2: LinearSet<_> = b.symmetric_difference(&a).cloned().collect();

assert_eq!(diff1, diff2);
assert_eq!(diff1, [1, 4].iter().cloned().collect());

fn intersection<'a>(&'a self, other: &'a LinearSet<T>) -> Intersection<'a, T>

Visit the values representing the intersection.

Examples

use linear_map::set::LinearSet;;
let a: LinearSet<_> = [1, 2, 3].iter().cloned().collect();
let b: LinearSet<_> = [4, 2, 3, 4].iter().cloned().collect();

// Print 2, 3 in arbitrary order.
for x in a.intersection(&b) {
    println!("{}", x);
}

let intersection: LinearSet<_> = a.intersection(&b).cloned().collect();
assert_eq!(intersection, [2, 3].iter().cloned().collect());

fn union<'a>(&'a self, other: &'a LinearSet<T>) -> Union<'a, T>

Visit the values representing the union.

Examples

use linear_map::set::LinearSet;;
let a: LinearSet<_> = [1, 2, 3].iter().cloned().collect();
let b: LinearSet<_> = [4, 2, 3, 4].iter().cloned().collect();

// Print 1, 2, 3, 4 in arbitrary order.
for x in a.union(&b) {
    println!("{}", x);
}

let union: LinearSet<_> = a.union(&b).cloned().collect();
assert_eq!(union, [1, 2, 3, 4].iter().cloned().collect());

fn len(&self) -> usize

Returns the number of elements in the set.

Examples

use linear_map::set::LinearSet;;

let mut v = LinearSet::new();
assert_eq!(v.len(), 0);
v.insert(1);
assert_eq!(v.len(), 1);

fn is_empty(&self) -> bool

Returns true if the set contains no elements.

Examples

use linear_map::set::LinearSet;;

let mut v = LinearSet::new();
assert!(v.is_empty());
v.insert(1);
assert!(!v.is_empty());

fn drain(&mut self) -> Drain<T>

Clears the set, returning all elements in an iterator.

fn clear(&mut self)

Clears the set, removing all values.

Examples

use linear_map::set::LinearSet;;

let mut v = LinearSet::new();
v.insert(1);
v.clear();
assert!(v.is_empty());

fn contains<Q: ?Sized>(&self, value: &Q) -> bool where
T: Borrow<Q>,
Q: Eq,

Returns true if the set contains a value.

The value may be any borrowed form of the set's value type, but Eq on the borrowed form must match those for the value type.

Examples

use linear_map::set::LinearSet;;

let set: LinearSet<_> = [1, 2, 3].iter().cloned().collect();
assert_eq!(set.contains(&1), true);
assert_eq!(set.contains(&4), false);

fn is_disjoint(&self, other: &LinearSet<T>) -> bool

Returns true if the set has no elements in common with other. This is equivalent to checking for an empty intersection.

Examples

use linear_map::set::LinearSet;;

let a: LinearSet<_> = [1, 2, 3].iter().cloned().collect();
let mut b = LinearSet::new();

assert_eq!(a.is_disjoint(&b), true);
b.insert(4);
assert_eq!(a.is_disjoint(&b), true);
b.insert(1);
assert_eq!(a.is_disjoint(&b), false);

fn is_subset(&self, other: &LinearSet<T>) -> bool

Returns true if the set is a subset of another.

Examples

use linear_map::set::LinearSet;;

let sup: LinearSet<_> = [1, 2, 3].iter().cloned().collect();
let mut set = LinearSet::new();

assert_eq!(set.is_subset(&sup), true);
set.insert(2);
assert_eq!(set.is_subset(&sup), true);
set.insert(4);
assert_eq!(set.is_subset(&sup), false);

fn is_superset(&self, other: &LinearSet<T>) -> bool

Returns true if the set is a superset of another.

Examples

use linear_map::set::LinearSet;;

let sub: LinearSet<_> = [1, 2].iter().cloned().collect();
let mut set = LinearSet::new();

assert_eq!(set.is_superset(&sub), false);

set.insert(0);
set.insert(1);
assert_eq!(set.is_superset(&sub), false);

set.insert(2);
assert_eq!(set.is_superset(&sub), true);

fn insert(&mut self, value: T) -> bool

Adds a value to the set.

If the set did not have a value present, true is returned.

If the set did have this key present, false is returned.

Examples

use linear_map::set::LinearSet;;

let mut set = LinearSet::new();

assert_eq!(set.insert(2), true);
assert_eq!(set.insert(2), false);
assert_eq!(set.len(), 1);

fn remove<Q: ?Sized>(&mut self, value: &Q) -> bool where
T: Borrow<Q>,
Q: Eq,

Removes a value from the set. Returns true if the value was present in the set.

The value may be any borrowed form of the set's value type, but Eq on the borrowed form must match those for the value type.

Examples

use linear_map::set::LinearSet;;

let mut set = LinearSet::new();

set.insert(2);
assert_eq!(set.remove(&2), true);
assert_eq!(set.remove(&2), false);